Revert "Merge branch 'main' of https://github.com/ammaraftabkhan/Accounts"
This reverts commit c19026b70242bb0f1bfc24bc3333187fc6f59f54, reversing
changes made to 050cf5f11a00b5d1c47debf120e91c70f641913a.

+++ Accounts.Common/Virtual Models/VM_CivilEntity.cs
@@ -9,10 +9,10 @@ namespace Accounts.Common.Virtual_Models
 {
     public class VM_CivilEntity
     {
-        public long CivilEntityId { get; set; }
+        public int CivilEntityId { get; set; }
         public string? CivilEntityName { get; set; }
         public int CivilLevelId { get; set; }
-        public long CivilParentId { get; set; }
+        public int CivilParentId { get; set; }
         public string? FlagImage { get; set; }
         [Column(TypeName = "datetime")]
         public DateTime CreatedOn { get; set; }
